requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
sqlalchemy
psycopg2-binary
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
python-dotenv==1.0.0
pytest==7.4.3
httpx==0.25.1